+++ internal/broadcast/syshandler_datatype_test.go
@@ -29,40 +29,6 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
-func TestHandleSystemBroadcastUnknown(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: "uknown",
-		},
-	}, []*fftypes.Data{})
-	assert.False(t, valid)
-	assert.NoError(t, err)
-}
-
-func TestGetSystemBroadcastPayloadMissingData(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-	valid := bm.getSystemBroadcastPayload(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: "uknown",
-		},
-	}, []*fftypes.Data{}, nil)
-	assert.False(t, valid)
-}
-
-func TestGetSystemBroadcastPayloadBadJSON(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-	valid := bm.getSystemBroadcastPayload(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: "uknown",
-		},
-	}, []*fftypes.Data{}, nil)
-	assert.False(t, valid)
-}
-
 func TestHandleSystemBroadcastDatatypeOk(t *testing.T) {
 	bm, cancel := newTestBroadcast(t)
 	defer cancel()
@@ -289,165 +255,5 @@ func TestHandleSystemBroadcastDatatypeDuplicate(t *testing.T) {
 	assert.NoError(t, err)
 
 	mdm.AssertExpectations(t)
-	mbi.AssertExpectations(t)
-}
-
-func TestHandleSystemBroadcastNSOk(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	ns := &fftypes.Namespace{
-		ID:   fftypes.NewUUID(),
-		Name: "ns1",
-	}
-	b, err := json.Marshal(&ns)
-	assert.NoError(t, err)
-	data := &fftypes.Data{
-		Value: fftypes.Byteable(b),
-	}
-
-	mbi := bm.database.(*databasemocks.Plugin)
-	mbi.On("GetNamespace", mock.Anything, "ns1").Return(nil, nil)
-	mbi.On("UpsertNamespace", mock.Anything, mock.Anything, true).Return(nil)
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{data})
-	assert.True(t, valid)
-	assert.NoError(t, err)
-
-	mbi.AssertExpectations(t)
-}
-
-func TestHandleSystemBroadcastNSUpsertFail(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	ns := &fftypes.Namespace{
-		ID:   fftypes.NewUUID(),
-		Name: "ns1",
-	}
-	b, err := json.Marshal(&ns)
-	assert.NoError(t, err)
-	data := &fftypes.Data{
-		Value: fftypes.Byteable(b),
-	}
-
-	mbi := bm.database.(*databasemocks.Plugin)
-	mbi.On("GetNamespace", mock.Anything, "ns1").Return(nil, nil)
-	mbi.On("UpsertNamespace", mock.Anything, mock.Anything, true).Return(fmt.Errorf("pop"))
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{data})
-	assert.False(t, valid)
-	assert.EqualError(t, err, "pop")
-
-	mbi.AssertExpectations(t)
-}
-
-func TestHandleSystemBroadcastNSMissingData(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{})
-	assert.False(t, valid)
-	assert.NoError(t, err)
-}
-
-func TestHandleSystemBroadcastNSBadID(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	ns := &fftypes.Namespace{}
-	b, err := json.Marshal(&ns)
-	assert.NoError(t, err)
-	data := &fftypes.Data{
-		Value: fftypes.Byteable(b),
-	}
-
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{data})
-	assert.False(t, valid)
-	assert.NoError(t, err)
-}
-
-func TestHandleSystemBroadcastNSBadData(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	data := &fftypes.Data{
-		Value: fftypes.Byteable(`!{json`),
-	}
-
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{data})
-	assert.False(t, valid)
-	assert.NoError(t, err)
-}
-
-func TestHandleSystemBroadcastDuplicate(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	ns := &fftypes.Namespace{
-		ID:   fftypes.NewUUID(),
-		Name: "ns1",
-	}
-	b, err := json.Marshal(&ns)
-	assert.NoError(t, err)
-	data := &fftypes.Data{
-		Value: fftypes.Byteable(b),
-	}
-
-	mbi := bm.database.(*databasemocks.Plugin)
-	mbi.On("GetNamespace", mock.Anything, "ns1").Return(ns, nil)
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{data})
-	assert.False(t, valid)
-	assert.NoError(t, err)
-
-	mbi.AssertExpectations(t)
-}
-
-func TestHandleSystemBroadcastDupCheckFail(t *testing.T) {
-	bm, cancel := newTestBroadcast(t)
-	defer cancel()
-
-	ns := &fftypes.Namespace{
-		ID:   fftypes.NewUUID(),
-		Name: "ns1",
-	}
-	b, err := json.Marshal(&ns)
-	assert.NoError(t, err)
-	data := &fftypes.Data{
-		Value: fftypes.Byteable(b),
-	}
-
-	mbi := bm.database.(*databasemocks.Plugin)
-	mbi.On("GetNamespace", mock.Anything, "ns1").Return(nil, fmt.Errorf("pop"))
-	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
-		Header: fftypes.MessageHeader{
-			Topic: fftypes.SystemTopicBroadcastNamespace,
-		},
-	}, []*fftypes.Data{data})
-	assert.False(t, valid)
-	assert.EqualError(t, err, "pop")
-
 	mbi.AssertExpectations(t)
 }
