+++ internal/broadcast/syshandler_test.go
@@ -1,0 +1,59 @@
+// Copyright © 2021 Kaleido, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package broadcast
+
+import (
+	"context"
+	"testing"
+
+	"github.com/kaleido-io/firefly/pkg/fftypes"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestHandleSystemBroadcastUnknown(t *testing.T) {
+	bm, cancel := newTestBroadcast(t)
+	defer cancel()
+	valid, err := bm.HandleSystemBroadcast(context.Background(), &fftypes.Message{
+		Header: fftypes.MessageHeader{
+			Topic: "uknown",
+		},
+	}, []*fftypes.Data{})
+	assert.False(t, valid)
+	assert.NoError(t, err)
+}
+
+func TestGetSystemBroadcastPayloadMissingData(t *testing.T) {
+	bm, cancel := newTestBroadcast(t)
+	defer cancel()
+	valid := bm.getSystemBroadcastPayload(context.Background(), &fftypes.Message{
+		Header: fftypes.MessageHeader{
+			Topic: "uknown",
+		},
+	}, []*fftypes.Data{}, nil)
+	assert.False(t, valid)
+}
+
+func TestGetSystemBroadcastPayloadBadJSON(t *testing.T) {
+	bm, cancel := newTestBroadcast(t)
+	defer cancel()
+	valid := bm.getSystemBroadcastPayload(context.Background(), &fftypes.Message{
+		Header: fftypes.MessageHeader{
+			Topic: "uknown",
+		},
+	}, []*fftypes.Data{}, nil)
+	assert.False(t, valid)
+}
